pyproject.toml
[project]
name = "realtime-analytics-fastapi"
version = "0.1.0"
description = "Realtime analytics dashboard"
authors = [
    {name = "Данил Тиводар",email = "danilkativodar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "h11 (>=0.16.0,<0.17.0)",
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg (>=3.2.10,<4.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "celery[flower] (>=5.5.3,<6.0.0)",
    "aiohttp (>=3.13.1,<4.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "fastapi-users[sqlalchemy] (>=15.0.1,<16.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "aiosqlite (>=0.22.0,<0.23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.2.0"
flake8 = "^7.3.0"
black = "^25.9.0"
isort = "^6.1.0"
httpx = "^0.28.1"
asgi-lifespan = "^2.1.0"
nest-asyncio = "^1.6.0"

